% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringPlots.R
\name{assignLabelSpecs}
\alias{assignLabelSpecs}
\title{Replicate color mapping with point types and assign legend point types}
\usage{
assignLabelSpecs(extraParams_ls, sampClasses_v, uSampClasses_v, normmat)
}
\arguments{
\item{extraParams_ls}{list of extra plotting parameters. Name of element must be exact name of default plotting argument}

\item{sampClasses_v}{vector of length == ncol(normmat) containing grouping identifier for each sample}

\item{uSampClasses_v}{vector of each unique grouping identifier}

\item{normmat}{data matrix, with unique & informative names}
}
\value{
list of vectors of plotting parameters
}
\description{
This function is called from within makeMDSplot, appropriately assigns point size for plot and legend.
}
