% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringPlots.R
\name{assignMappingSpecs}
\alias{assignMappingSpecs}
\title{Map colors to plotting factor and samples}
\usage{
assignMappingSpecs(attribs, oneclass, colorspec, varPoints_v = NULL)
}
\arguments{
\item{attribs}{list of sample classifications to be tracked in clustering. Each list element contains a string vector with one lable per sample.}

\item{oneclass}{string name of attribs element to be used in MDS plot}

\item{colorspec}{vector of color specifiers for colorRampPalette. colors are generated per sample by their clustering variable values}

\item{varPoints_v}{vector of values to scale points by. Default is no scaling}
}
\value{
list of vectors of plotting paramters
}
\description{
This function is called from within makeMDSplot, appropriately colors vectors and changes point size based on input
}
