% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringPlots.R
\name{createLegend}
\alias{createLegend}
\title{Create legend to display on plots}
\usage{
createLegend(legendPos_v, uSampClasses_v, uPlotColors_v, pchLegend_v)
}
\arguments{
\item{legendPos_v}{"auto" means legend in default configuration at upper limit of user space, 
can specify standard options (see ?legend)}

\item{uSampClasses_v}{vector of each unique grouping identifier}

\item{uPlotcolors_v:}{vector of colors, one for each unique entry in uSampClasses_v
(should be created using colorRampPalette - see assignMappingSpecs)}

\item{pchLegend_v:}{vector of integers specifying point types. Should be either length of 1 or
length(pchLegend_v) == length(uSampClasses_v)}
}
\value{
legend object
}
\description{
This function is currently called from within makeMDSplot, although it may be suitable to other
functions as well, as-is or with slight tweaking.
}
