% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringPlots.R
\name{gene_pcaplot}
\alias{gene_pcaplot}
\title{Make a PCA plot}
\usage{
gene_pcaplot(exprdat, sampleid, groupdat = NULL, colorfactor = NULL,
  shapefactor = NULL, plot_sampleids = TRUE, pcnum = 1:2,
  plottitle = "PCA Plot")
}
\arguments{
\item{exprdat}{matrix of expression data}

\item{sampleid}{vector of sample IDs}

\item{groupdat}{matrix of groups of samples as columns}

\item{colorfactor}{vector to use in coloring points}

\item{shapefactor}{vector to use in changing point shapes}

\item{plot_sampleids}{logical. if TRUE (default) plot a legend with sample IDs}

\item{pcnum}{vector of pcnum attributes to extract for axis labels}

\item{plottitle}{vector of title for plot}
}
\value{
PCA plot
}
\description{
borrowed from Jessica Minnier, 2016. adapted from DESeq2:::plotPCA.DESeqTransform
}
