% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringPlots.R
\name{getCidFromHeatmap}
\alias{getCidFromHeatmap}
\title{Return a list of cluster IDs and elements of dendrogram in left-right order}
\usage{
getCidFromHeatmap(aheatmapObj, numSubTrees, cutByRowLogic = T)
}
\arguments{
\item{aheatmapObj}{object returned by the aheatmap() function from the NMF package}

\item{numSubTrees}{numerical vector of length 1 specifying number of subclades/subtrees to split the dendrogram row (or column) into.}

\item{cutByRowLogic}{boolean specifying to return cluster IDs and orders for rows (TRUE) or columns (FALSE)}
}
\description{
Takes a heatmap object (see makeHeatmap), and splits rows (or columns) into a dendrogram with specified number of trees. Cluster IDs and elemenets returned.
}
