% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringPlots.R
\name{makeHeatmap}
\alias{makeHeatmap}
\title{Make heatmap of ratio data}
\usage{
makeHeatmap(ratiomat, attribs, plottitle, subtitle = NULL, normmat = NULL,
  clim.pct = 0.99, clim_fix = NULL, colorbrew = "-PiYG:64",
  cexRow = 1e-05, annRow = NA, annColors = NA, cexCol = min(0.2 +
  1/log10(ncol(ratiomat)), 1.2), labcoltype = c("colnames", "colnums"),
  labRowType = NULL, setCol_v = NULL, setRowv = NULL, colOrder_v = NULL,
  inclLegend = TRUE)
}
\arguments{
\item{ratiomat}{ratio data matrix, optimally derived from normmat,
with unique & informative colnames}

\item{attribs}{list of sample classifications to be tracked in clustering
each list element contains a string vector with one label per sample
set to NA to omit}

\item{plottitle}{title for all plots}

\item{normmat}{abundance data matrix, optionally used to set color limits
set to null to use ratiomat for color limits}

\item{clim.pct}{0:1 - fraction of data to limit max color}

\item{clim_fix}{if set, max abs ratio to show; data>clim_fix are shown as clim_fix}

\item{colorbrew}{is a colorBrewer string specifying a heatmap color scale 
colors are generated per sample by their clustering variable values}

\item{cexRow}{rowlabel size for heatmap, set to ~invisible by default}

\item{annRow}{named lists of row annotations; can be single list}

\item{annColors}{named lists of annotation colors; names should match names in annRow and attribs}

\item{cexCol}{collabel size for heatmap, set to aheatmap default by default}

\item{labcoltype}{colnames to show ratiomat column names, colnums to show col #s}

\item{labRowType}{labels for rows; default NULL doesn't label}

\item{setCol_v}{name of attrib to sort matrix columns by (will turn off column-clustering in heatmap output)
Defaults to NULL to keep original clustering (hierarchical)}

\item{setRowv}{pass through to Rowv argument of aheatmap. Default to NULL, hierarchical clustering}

\item{colOrder_v}{Vector containing all unique values of attribs[[setColv]] in desired output order. Defaults
to NULL so that if setColv is specified alone, output order will be alphabetical.}

\item{inclLegend}{logical, should legend be included in plot.}
}
\description{
This function makes a heatmap of ratio data, displaying experimental design values as tracks
Uses the matrix values to cluster the data
}
