% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringPlots.R
\name{makeMDSplot}
\alias{makeMDSplot}
\title{Make multi-dimensional scaling plot}
\usage{
makeMDSplot(normmat, attribs, oneclass, colorspec, plottitle, subtitle = NULL,
  ngenes = NULL, legendPos_v = "auto", varPoints_v = NULL, ...)
}
\arguments{
\item{normmat}{data matrix, with unique & informative colnames}

\item{attribs}{list of sample classifications to be tracked in clustering
each list element contains a string vector with one label per sample}

\item{oneclass}{string name of attribs element to be used in MDS plot}

\item{colorspec}{vector of color specifiers for colorRampPalette
colors are generated per sample by their clustering variable values}

\item{plottitle}{title for all plots}

\item{subtitle}{optional subtitle to add below title on this plot}

\item{ngenes}{number of "top" genes for plotMDS to select for distance calculation}

\item{legendPos_v}{"auto" means legend in default configuration at upper limit of user space, can specify
standard options (see ?legend)}

\item{varPoints_v}{vector of values to scale points by. Default is no scaling. 
If used, values are placed into 15 bins and assigned cex values from 0.5 to 1.9}

\item{...}{any extra plotting parameters to pass to plotMDS. (change plotting points, axis labels, etc.)}
}
\value{
mds plot
}
\description{
This function considers only a single clustering variable in coloring MDS plot... Will need to be gneralized
Uses the matrix values to cluster the data
}
