% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryPlots.R
\name{makePairwisePlotMatrix}
\alias{makePairwisePlotMatrix}
\title{Wrapper for ggpairs, which generates a matrix of plots with a given data set}
\usage{
makePairwisePlotMatrix(norm_mat, plotdata, plot2file = FALSE, reso = 600,
  ...)
}
\arguments{
\item{...}{addition parameters that ggpairs() can take
normmat is a matrix of experimental data in columns (with headers!)
attribs is a vector of experimental categories (one entry per column in normmat)
plotdata is a list of info relevant to labeling and saving the plot
 plotdir:  plot destination directory
 plotbase:  base filename for the plot
 plottitle:  title for all plots
 @return 
 @examples
 normInst_mat = summ_ls$loess
 plotdataInst = list(plotdir = "~/Desktop/", plotbase="BaseTest", plottitle = "titleTest")
 plot2fileInst = TRUE
 makePairwisePlotMatrix(norm_mat = normInst_mat, plotdata = plotdataInst, plot2file=plot2fileInst, pch=".")}
}
\description{
This is intended to show individual replicates vs. average of experimental group
It will create one plot per column, and an average vs average
}
