% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryPlots.R
\name{maplot}
\alias{maplot}
\title{Generates Bland-Altman plot: log ratio vs. average of two samples}
\usage{
maplot(v1, v2, v12names = list(v1 = "Sample 1", v2 = "Sample 2"), plotdata,
  yrange = NULL, plot2file = FALSE, plotIDOffset = 0)
}
\arguments{
\item{v1}{a vector of abundance data for a sample of interest}

\item{v2}{a second vector of same length as v1.}

\item{v12names}{a list with two elements: v1 and v2, which are strings describing vectors v1 and v2. These strings will appear in the plot. Defaults to calling them "Sample 1" and "Sample 2".}

\item{plotdata}{a list with three elements of info relevant to labeling and saving the plot: 1. plotdir-plot destination directory. 2. plotbase-base filename for the plot. 3. plottitle-title for all plots}

\item{yrange}{a numerical vector of length 2 that sets the y-axis range. Default is NULL.}

\item{plot2file}{a boolean specifying whether the plot is to be saved to file according to the file-naming specification laid out in plotdata. Default is FALSE.}

\item{plotIDOffset}{a number the specifies offset value for the plot ID. Default value is 0.}
}
\value{
#Attn. #Feedback
}
\description{
Note:  fivenumber or quantile vector will be included in the y-axis label, so it will be clear if the plot has been limited. 
... #Attn. #Feedback flesh out description more
}
\examples{
norm_mat = summ_ls$loess
v1Inst = norm_mat[,1]
v2Inst = norm_mat[,2]
v12namesInst = list(v1=gsub('(^.+?)_.*','\\\\1', colnames(norm_mat)[1]),v2=gsub('(^.+?)_.*','\\\\1', colnames(norm_mat)[2]))
plotdataInst = list(plotdir = "~/Desktop/", plotbase ="MA", plottitle = "MATestPlot")
yrangeInst = NULL
plot2fileInst = TRUE
plotIDOffsetInst = 0
maplot(v1=v1Inst, v2=v2Inst, v12names = v12namesInst, plotdata = plotdataInst, yrange = yrangeInst, plot2file = plot2fileInst, plotIDOffset = plotIDOffsetInst)
}
