% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryPlots.R
\name{plot_raw_v_norm}
\alias{plot_raw_v_norm}
\title{Plots raw data against normalized data}
\usage{
plot_raw_v_norm(linear_scale_raw_mat_with_rownames, str_of_raw_name = NULL,
  log2_scale_norm_mat_with_rownames, str_of_norm_name = NULL,
  xrng_vec = range(range(linear_scale_raw_mat_with_rownames, na.rm = T),
  range(2^log2_scale_norm_mat_with_rownames, na.rm = T)),
  yrng_vec = log2(xrng_vec), save_as_png = F,
  string_to_lead_file_name_with = NULL, color_by_vec_of_str = F,
  color_vec_of_strings = NULL)
}
\arguments{
\item{linear_scale_raw_mat_with_rownames}{Matrix of data}

\item{str_of_raw_name}{Character vector of name of raw data}

\item{log2_scale_norm_mat_with_rownames}{Matrix of data}

\item{str_of_norm_name}{Character vector of name of norm data}

\item{xrng_vec}{place holder}

\item{yrng_vec}{place holder}

\item{save_as_png}{logical. Default is FALSE. If TRUE, write plot to png file.}

\item{string_to_lead_file_name_with}{Vector with base portion of output file names}

\item{color_vec_of_strings}{vector of color specifications that can be used for plotting colors}

\item{color_by_vec_of_string}{logical. Default is FALSE. If TRUE, apply color specifications to samples}
}
\description{
Need to update the parameters
}
