% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringPlots.R
\name{plotRatios}
\alias{plotRatios}
\title{Differential expression visualizations. Heatmap and MDS plot of ratios}
\usage{
plotRatios(ratiomat, attribs, oneclass, plotdata, colorspec, rowmask = NULL,
  tag = "selected", NAtol = ncol(ratiomat)/2, SDtol = 0.01,
  heatmap_plot = TRUE, MDS_plot = TRUE, clim.pct = 0.99,
  clim_fix = NULL, plot2file = FALSE, filesep = "/", annColors = NA,
  annRow = NA, cexRow = 1e-05, png_res = 300)
}
\arguments{
\item{ratiomat}{matrix of ratio data in columns (with headers!)}

\item{attribs}{list of sample classifications to be tracked in clustering
each list element contains a string vector with one label per sample}

\item{oneclass}{string name of attribs element to be used in MDS plot}

\item{plotdata}{list of info relevant to labeling and saving the plot
plotdir  plot destination directory
plotbase  base filename for the plot
plottitle  title for all plots}

\item{colorspec}{is a vector of color specifiers for colorRampPalette}

\item{rowmask}{optional mask of rows to plot
default is to create an all-TRUE rowmask of length nrow(ratiomat)}

\item{tag}{word or phrase to indicate what's special about _these_ ratios}

\item{NAtol}{max number of NAs tolerated; rows with more are masked out}

\item{SDtol}{min SD within row tolerated; rows with less are masked out}

\item{heatmap_plot}{boolean to plot (TRUE) or not (FALSE) heatmap}

\item{MDS_plot}{boolean to plot (TRUE) or not (FALSE) MDS plot}

\item{clim.pct}{0:1 - fraction of data to limit max color}

\item{clim_fix}{if set, max abs ratio to show; data>clim_fix are shown as clim_fix}

\item{annColors}{optional named lists of annotation colors;
names should match names in annRow and attribs}

\item{annRow}{optional named lists of row annotations; can be single list}

\item{cexRow}{rowlabel size for heatmap, set to ~invisible by default}

\item{png_res}{resolution of saved png files in dpi; default 300}
}
\description{
This function takes a matrix of ratios and displays them graphically in a heatmap and/or MDS plot
}
