% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleTestingCorrection.R
\name{qc.clusters}
\alias{qc.clusters}
\title{Creates a heatmap and an MDS plot}
\usage{
qc.clusters(normmat, rawmat, attribs, oneclass, plotdata, colorspec,
  center = "norm", clim.pct, mask = NULL, plot2file = FALSE,
  filesep = "/")
}
\arguments{
\item{normmat}{data matrix}

\item{rawmat}{raw data matrix (optionally used to center normmat)}

\item{attribs}{list of sample classifications to be tracked in clustering 
each list element contains a string vector with one label per sample}

\item{oneclass}{string name of attribs element to be used in MDS plot}

\item{plotdata}{is a list of info relevant to labeling and saving the plot
plotdir:  plot destination directory
plotbase:  base filename for the plot
plottitle:  title for all plots}

\item{colorspec}{is a vector of color specifiers for colorRampPalette}

\item{center}{center data on 'norm' for normalized average
none indicates no-centering needed for zero-centered data
any other value will trigger centering on the average of rawmat}

\item{clim.pct}{0:1 - fraction of data to limit max color}

\item{mask}{NULL default masks rows with SD > SD of entire data set; 
Numeric vector of length 1 masks rows with SD > value
Vector of length nrows(normmat) is a custom mask   ***** TBImplemented}

\item{plot2file}{if true, plot to file as specified in plotdata}
}
\description{
This function currently only considers a single clustering variable and will need to be generalized.
Data will be centered around the average of all data in the normalized data set
As a QC plot, want to filter out non-changing data, otherwise it takes a LONG time to complete
}
