% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleTestingCorrection.R
\name{qcQ_wrapper}
\alias{qcQ_wrapper}
\title{Wrapper for qcQvalue function to verify q-value analysis terms}
\usage{
qcQ_wrapper(norm_x, p_mat, q_list, metadata, plotdata, plot2file = TRUE,
  bonusMDS = FALSE, oneclass = NULL, qcut = 0.1, histbins = 40,
  colorspec = c("#7b3294", "#008837"), facSel = NULL, filesep = "/")
}
\arguments{
\item{norm_x}{abundance data input into regression (eg, regressMatrix()). nrow(norm_x)==length(p_mat)}

\item{p_mat}{matrix of p-values for each factor. Returned by regressMatrix()}

\item{q_list}{S3 object; list of p-value lists for each factor. Returned by regressMatrix()}

\item{metadata}{list of factors; each list element is a factor (eg, sex = c("m", "f"...)) #Feedback perhaps as.factor(c("m","f"))?
Vector order same as norm_x column order
was attribs previously}

\item{plotdata}{is a list of info relevant to labeling and saving the plot With the following elements:
plotdir:  plot destination directory
plotbase:  base filename for the plot. Suggest: bias reduction method
plottitle:  title for all plots
plotSubtitle:  optional subtitle; suggest lm_expression}

\item{plot2file}{logical. if TRUE, write files using plotdata, otherwise to studio graphics viewer}

\item{bonusMDS}{logical. if TRUE plots an MDS of data at qcut (see below). Default is FALSE (no plot)}

\item{oneclass}{metadata vector for labeling bonusMDS data; if NULL or not specified, defaults to the first element of metadata}

\item{qcut}{either a number between 0 and 1 used as an upper qvalue limit for MDS, OR a number >1, assumed to be the top n qvalues to use in MDS}

\item{histbins}{number of bins for p-value histogram; may want to change if there is suspicious behavior}

\item{colorspec}{optional range of colors for MDS plot. Defaults to hotpink/green!}

\item{facSel}{optional vector of experimental factors to plot (rather than all factors) #Feedback to plot for the MDS plot? What if you want to plot based on second element of metadata list?}

\item{filesep}{string that adds to end of directory path.}
}
\value{
Writes plots to files or to Rstudio viewer. Notes - Looking for even descent to pi0 in top left tuning plot. The slow/steep rise in q-values
in remaining plots depends on the resolving power of the data.
}
\description{
Creates plots for each design factor (default behavior), or for factors facSel
1) Histogram of p-values that were included in the design
2) qvalue's default plots with full q-value range (c(0,1)).
3) Optional MDS plot at q-value cut
}
