% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleTestingCorrection.R
\name{qcQvalues}
\alias{qcQvalues}
\title{Checks proper running of q-value analysis}
\usage{
qcQvalues(norm_x, pvalue_v, obj_qvalue, attribs, oneclass, plotdata, colorspec,
  qcut = 1, histbins = 40, plot2file = FALSE, p_hist = TRUE,
  q_plots = TRUE, MDS_plot = FALSE, filesep = "/")
}
\arguments{
\item{norm_x}{abundance data input into regression (eg, regressMatrix()). nrow(norm_x)==length(p_mat)}

\item{pvalue_v}{vector of p-values previously used as input to qvalue()}

\item{attribs}{list of sample classifications to be tracked in clustering
each list element contains a string vector with one label per sample}

\item{oneclass}{metadata vector for labeling bonusMDS data; if NULL or not specified, defaults to the first element of metadata}

\item{plotdata}{is a list of info relevant to labeling and saving the plot With the following elements:
plotdir:  plot destination directory
plotbase:  base filename for the plot. Suggest: bias reduction method
plottitle:  title for all plots
plotSubtitle:  optional subtitle; suggest lm_expression}

\item{colorspec}{optional range of colors for MDS plot. Defaults to hotpink/green!}

\item{qcut}{either a number between 0 and 1 used as an upper qvalue limit for MDS,
OR a number >1, assumed to be the top n qvalues to use in MDS}

\item{histbins}{number of bins for p-value histogram; may want to change if there is suspicious behavior}

\item{plot2file}{logical. if TRUE, write files using plotdata, otherwise to studio graphics viewer}

\item{p_hist}{logical. TRUE (default) to plot this type.}

\item{q_plots}{logical. TRUE (default) to plot this type.}

\item{MDS_plot}{logical. TRUE to plot this type. FALSE is default. Will plot after selecting q-value cut based on QC plots. Use plotRatios() instead!}

\item{filesep}{string that adds to end of directory path.}

\item{obj_qvalue:}{S3 object of qvalue class (a list!) returned by qvalue()}
}
\value{
Writes plots to files or to Rstudio viewer. Notes - Looking for even descent to pi0 in top left tuning plot. The slow/steep rise in q-values
in remaining plots depends on the resolving power of the data.
}
\description{
Creates plots for each design factor (default behavior), or for factors facSel
1) Histogram of p-values that were included in the design. Must not have right peak.
2) qvalue's default plots with full q-value range (c(0,1)).
3) Optional MDS plot at q-value cut
}
