% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryPlots.R
\name{scatterplot}
\alias{scatterplot}
\title{Display individual replicates vs. average of experimental group in a scatterplot}
\usage{
scatterplot(normmat, attribs, plotdata, plot2file = FALSE, plotIDOffset = 0)
}
\arguments{
\item{normmat}{matrix of exerimental data. Rows = observations, columns = samples/experiments (requires named rows and columns)}

\item{attribs}{vector of experimental categories of same length as ncol(normmat)}

\item{plotdata}{list of vectors relevant to labeling and saving plot
plotdir:  plot destination directory
plotbase:  base filename for the plot
plottitle:  title for all plots}
}
\description{
Uses matrix of experimental observations from multiple testing groups. 
Compares individual observations to corresponding group average.
Creates one plot per column, and an average vs average
}
