% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringPlots.R
\name{setColSpecs}
\alias{setColSpecs}
\title{Set column specs for heatmap}
\usage{
setColSpecs(ratiomat, attribs, setCol_v = NULL, colOrder_v = NULL,
  labcoltype = c("colnames", "colnums"))
}
\arguments{
\item{ratiomat}{ratio data matrix, optimally derived from normmat,
with unique & informative colnames}

\item{attribs}{list of sample classifications to be tracked in clustering
each list element contains a string vector with one label per sample
set to NA to omit}

\item{setCol_v}{name of attrib to sort matrix columns by (will turn off column-clustering in heatmap output)
Defaults to NULL to keep original clustering (hierarchical)}

\item{colOrder_v}{Vector containing all unique values of attribs[[setColv]] in desired output order. Defaults
to NULL so that if setColv is specified alone, output order will be alphabetical.}

\item{labcoltype}{colnames to show ratiomat column names, colnums to show col #s}
}
\value{
list of column parameters
}
\description{
This function is called in makeHeatmap and determines column ordering and labelling
}
