% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryPlots.R
\name{summary.plots}
\alias{summary.plots}
\title{A wrapper for several plotting functions that help summarize abundance data #Feedback since rawmat in this case is on log2 scale, will not yet accommodate methylation data}
\usage{
\method{summary}{plots}(rawmat, normmat, mynorm, samp.labels, samp.classes,
  colorspec, plotdata, plot2file = FALSE, histbins = 40, expand.2D = 5,
  SDrange = 7, filesep = "/", plotIDOffset = 0, verbose = FALSE,
  BaseSample = NULL, yrange = NULL, MAplotOpt = FALSE,
  whichPlots_v = c("box", "scatter", "density", "spread"))
}
\arguments{
\item{rawmat}{a matrix of raw data. Should have minimal background addition and be scaled according to normalized data. In other words, rawmat should be the alograw slot of a normMatrix() call.}

\item{normmat}{a matrix of experimental data in columns (with headers!)}

\item{mynorm}{a label for the current normalization method #Feedback not sure what this data type is}

\item{samp.labels}{a vector of brief, pithy display labels for each sample}

\item{samp.classes}{a vector of tags for each sample, used to pick colors
a unique tag per sample colors by sample (e.g., c("01", "02", "03", "04"))
a shared tag by experimental groupings colors by experimental group (e.g. c("male", "female", "female", "male"))}

\item{colorspec}{a vector of color specifiers for colorRampPalette}

\item{plotdata}{a list of info relevant to labeling and saving the plot. Consists of three elements:
plotdir:  plot destination directory
plotbase:  base filename for the plot
plottitle:  title for all plots}

\item{histbins}{default number of bins for density plot}

\item{expand.2D}{a numeric multiplier of histbins for 2D histogram}

\item{SDrange}{maximum SD to plot in spread plots}

\item{filesep}{a string to designate file separators #Feedback not sure}

\item{plotIDOffset}{a number the specifies offset value for the plot ID. Default value is 0.}

\item{BaseSample}{a string of the column/sample name that will be used as the sample to which other samples will be compared for MA plots.}

\item{yrange}{a numerical vector of length 2 that sets the y-axis range. Default is NULL.}

\item{MAplotOpt}{a boolean that specifies whether the user wants to generate MA plots in this call}
}
\value{
if verbose is TRUE, a list of ___
}
\description{
summary.plots currently wraps a boxplot (1), scatterplot (2), density plot (3), sd vs. intensity plot (4), and MA plot (5). The MA plot behavior plots all sample columns in a matrix vs. one BaseSample.
}
\examples{
Examples go here
}
